DefaultSteps class divided into four classes DefaultStepsTest class divided into four classes Imports optimized
