+++ src/test/java/ru/alfabank/other/AkitaPageTest.java
@@ -28,7 +28,7 @@ import ru.alfabank.StubScenario;
 import ru.alfabank.alfatest.cucumber.api.AkitaEnvironment;
 import ru.alfabank.alfatest.cucumber.api.AkitaPage;
 import ru.alfabank.alfatest.cucumber.api.AkitaScenario;
-import ru.alfabank.steps.DefaultSteps1;
+import ru.alfabank.steps.defaultSteps.WebPageSteps;
 
 import java.io.File;
 import java.util.LinkedList;
@@ -46,14 +46,14 @@ public class AkitaPageTest {
     public static void setup() {
         akitaPageMock = new AkitaPageMock();
         AkitaScenario akitaScenario = AkitaScenario.getInstance();
-        DefaultSteps1 ds = new DefaultSteps1();
+        WebPageSteps wps = new WebPageSteps();
         Scenario scenario = new StubScenario();
         akitaScenario.setEnvironment(new AkitaEnvironment(scenario));
         String inputFilePath = "src/test/resources/AkitaPageMock.html";
         String url = new File(inputFilePath).getAbsolutePath();
         akitaScenario.setVar("Page", "file://" + url);
         page = akitaScenario.getEnvironment().getPage("AkitaPageMock");
-        ds.goToSelectedPageByLinkFromPropertyFile("AkitaPageMock", akitaScenario.getVar("Page").toString());
+        wps.goToSelectedPageByLink("AkitaPageMock", akitaScenario.getVar("Page").toString());
     }
 
     @AfterClass
